Refactor bunny-cdn tests and fix custom domain guide
Split the 1328-line test/lib/bunny-cdn.test.ts into focused files:
- test/lib/bunny-cdn/config.test.ts (env var feature flags)
- test/lib/bunny-cdn/edge-script.test.ts (API operations)
- test/lib/bunny-cdn/domain.test.ts (DNS, subdomains, validation)
- test/lib/bunny-cdn/settings.test.ts (DB settings)

Removed pointless delegation tests that just verified a wrapper
calls the underlying function. Added a useful test for the edge
case where DNS cleanup is skipped when no record ID is returned.

Added stubFetchJson and stubFetchRecorder helpers to test-utils
to replace duplicated fetch-stubbing boilerplate.

Fixed the admin guide custom domain section: BUNNY_SCRIPT_ID is
also required (not just BUNNY_API_KEY), clarified what validation
actually does, and added a Q&A about handling validation failures.
